// src/types/ui.types.ts
export enum SortOrder {
  ASC = 'asc',
  DESC = 'desc',
}

export enum FilterType {
  ALL = 'all',
  RANKED = 'ranked',
  PRACTICE = 'practice',
}

export enum EmptyStateType {
  NO_TESTS = 'no-tests',
  NO_RESULTS = 'no-results',
}

export enum SortBy {
  DATE = 'date',
  ACCURACY = 'accuracy',
  APM = 'apm',
}

export enum StatusVariant {
  SUCCESS = 'success',
  DANGER = 'danger',
  WARNING = 'warning',
  INFO = 'info',
}

export enum BadgeSize {
  SM = 'sm',
  MD = 'md',
  LG = 'lg',
}

export enum ColorModeVariant {
  ICON = 'icon',
  TEXT = 'text',
}